Add service-type admin APIs and mail support
Introduce admin service-type management and email capabilities: add AdminServiceTypeController, AdminServiceTypeService, ServiceTypeRequest/Response DTOs, and repository queries to list/count active registrations; update resident service registration tests to use Integer IDs. Add MailService, wire it into AdminUserService, add spring-boot-starter-mail dependency, enable async processing, and configure mail properties & optional .env import in application.yaml. Also add an empty AdminReportController, minor controller formatting cleanups, update .gitignore and workspace settings, and include frontend/package-lock updates plus a new admin BuildingDetailPage and router changes.

+++ backend/src/main/java/com/apartmentmanagement/dto/request/resident/ServiceRegistrationRequest.java
@@ -2,10 +2,9 @@ package com.apartmentmanagement.dto.request.resident;
 
 import lombok.Data;
 
-
 @Data
 
 public class ServiceRegistrationRequest {
-    private Long serviceTypeId;
-    private String notes; 
+    private Integer serviceTypeId;
+    private String notes;
 }
